# swagger/swagger-v1.0.2.yaml
openapi: 3.0.3
info:
  title: Liga Nunez
  description: |- 
    HTTP requests y responses documentadas
    
    Links de utilidad:
    - [Requerimientos](https://docs.google.com/document/d/1_kFwMBbmoDiY1nCElglinTN5vLZckk4n5RgaN-1zyKg/edit?pli=1)
    - [Tablero JIRA](https://venedicto.atlassian.net/jira/software/projects/LGN/boards/9)
  version: 1.0.2
servers:
  - url: http://localhost:8080/api/ln/
...
paths:
  /user:
    get:
      tags:
        - user
      summary: Validación de existencia de usuario por email
      operationId: checkUserExistence
      parameters:
        - name: email
          in: header
          description: Email ingresado
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Usuario registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '400':
          description: Error en la request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '404':
          description: Usuario no registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '503':
          description: Base de datos/Servicio email no disponibles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
    post:
      tags:
        - user
      summary: Registro de un nuevo usuario
      operationId: createUser
      requestBody:
        description: Información del usuario (debe enviarse sin accessKey)
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '201':
          description: Usuario generado con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '400':
          description: Error en la request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '503':
          description: Base de datos/Servicio email no disponibles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
  /user/update/password:
    post:
      tags:
        - user
      summary: Solicitud de actualización de password
      operationId: requestUserPasswordUpdate
      parameters: []
      responses:
        '201':
          description: Solicitud de actualización de password realizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '404':
          description: Usuario inexistente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '503':
          description: Base de datos/Servicio email no disponibles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
    patch:
      tags:
        - user
      summary: Actualización de password
      operationId: updateUserPassword
      parameters:
        - name: authCode
          in: header
          description: Código recibido por correo
          required: false
          schema:
            type: string
        - name: newPassword
          in: header
          description: Nueva contraseña ingresada
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Contraseña actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '403':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '404':
          description: Usuario inexistente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '503':
          description: Base de datos no disponible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
  /user/update/permissions:
    patch:
      tags:
        - user
      summary: Actualización del nivel de permisos
      operationId: updateUserPermissionLevel
      parameters:
        - name: token
          in: header
          description: Token de autenticación del usuario
          required: false
          schema:
            type: string
        - name: permissionLevel
          in: header
          description: Nivel de permisos
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Permisos actualizados con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '403':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '404':
          description: Usuario inexistente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '503':
          description: Base de datos no disponible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
  /user/login:
    get:
      tags:
        - user
      summary: Ingreso de un usuario al sistema
      operationId: loginUser
      parameters:
        - name: email
          in: header
          description: Correo electrónico ingresado
          required: true
          schema:
            type: string
        - name: password
          in: header
          description: Contraseña ingresada
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Login exitoso (sin email ni accessKey)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLoginHttpResponse'
        '400':
          description: Error en la request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '503':
          description: Base de datos no disponible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
  /user/logout:
    get:
      tags:
        - user
      summary: Cierre de sesión de un usuario
      operationId: logoutUser
      parameters:
        - name: email
          in: header
          description: "Email del usuario"
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Sesión cerrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '400':
          description: Error en la request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '404':
          description: Usuario inexistente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
...
components:
  schemas:
...
    User:
      type: object
      properties:
        name:
          type: string
          example: "Juan Perez"
        age:
          type: integer
          example: 32
        address:
          type: string
          example: "Corrientes 800"
        permissionLevel:
          type: integer
          example: 1
        email:
          type: string
          example: "example@gmail.com"
        accessKey:
          type: string
          example: "943012"
    Error:
      type: object
      properties:
        code:
          type: string
          example: "P01"
        description:
          type: string
          example: "Error interno"
    HttpResponse:
      type: object
      properties:
        opCode:
          type: string
          example: "404"
        errors:
          type: array
          items:
            $ref: '#/components/schemas/Error'
    UserLoginHttpResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/HttpResponse'
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
        data:
          $ref: '#/components/schemas/User'
...